Use UInt64 for maxResponseSize

+++ SignalServiceKit/Backups/Archiving/BackupArchive+Constants.swift
@@ -9,6 +9,6 @@ extension BackupArchive {
 
     enum Constants {
         /// We reject downloading backup proto files larger than this.
-        static let maxDownloadSizeBytes: UInt = 100 * 1024 * 1024 // 100 MiB
+        static let maxDownloadSizeBytes: UInt64 = 100 * 1024 * 1024 // 100 MiB
     }
 }
